+++ components/helpers/applyFilter.ts
@@ -1,4 +1,5 @@
 interface DataObject {
+  name: string;
   [key: string]: any;
 }
 
@@ -8,6 +9,7 @@ interface FilterCriteria {
 
 interface Filter {
   [key: string]: string;
+  value: string;
 }
 
 interface FilterOption {
