+++ components/navigation/DocsNav.tsx
@@ -1,12 +1,41 @@
 import { useEffect, useState } from 'react';
 
-import type { DocsNavProps, SerializedBuckets  } from '@/types/navigation/DocsNav';
-
 import { buckets } from '../data/buckets';
 import DocsArrow from '../icons/DocsArrow';
 import IconHome from '../icons/Home';
 import DocsNavItem from './DocsNavItem';
 import SubCategoryDocsNav from './SubCategoryDocsNav';
+
+export interface Bucket {
+  name?: string;
+  title?: string;
+  description?: string;
+  link?: string;
+  className?: string;
+  borderClassName?: string;
+  Icon?: React.ComponentType<any>;
+  href?: string;
+  icon?: React.ComponentType<any> | null;
+}
+
+export interface SerializedBuckets {
+  [key: string]: Bucket;
+}
+
+export interface DocsNavProps {
+  item: {
+      children: {
+          [key: string]: any;
+      };
+      item: {
+          rootSectionId: string;
+          slug: string;
+          title: string;
+      };
+  };
+  active: string;
+  onClick?: () => void;
+}
 
 const serializedBuckets: SerializedBuckets = buckets.reduce((acc, bucket) => {
   acc[bucket.name || ''] = {
