--- types/navigation/DocsMobileMenu.ts
@@ -1,9 +1,0 @@
-export interface DocsMobileMenuProps {
-    post: {
-        slug: string;
-    };
-    navigation: {
-        [key: string]: any;
-    };
-    onClickClose?: () => void;
-}
