--- types/navigation/Filter.ts
@@ -1,11 +1,0 @@
-export interface Option {
-    value: string;
-    text: string;
-}
-
-export interface FilterProps {
-    data: any[];
-    onFilter: (data: any[]) => void;
-    checks: { name: string; options?: Option[] }[];
-    className: string;
-}
