--- types/navigation/MenuBlocks.ts
@@ -1,13 +1,0 @@
-export interface MenuItem {
-    title: string;
-    href: string;
-    description: string;
-    icon: React.ElementType;
-    className?: string;
-    comingSoon?: boolean;
-    beta?: boolean;
-}
-
-export interface MenuBlocksProps {
-    items?: MenuItem[];
-}
