--- types/navigation/SubCategoryDocsNav.ts
@@ -1,16 +1,0 @@
-export interface SubCategoryDocsNavProps {
-    subCategory: {
-      item: {
-        title: string;
-        slug: string;
-        href?: string;
-      };
-      children?: {
-        title: string;
-        slug: string;
-        href?: string;
-      }[];
-    };
-    activeItem: string;
-    onClick: () => void;
-  }
